+++ master/app/Http/Controllers/Admin/ImageProductController.php
@@ -12,6 +12,7 @@ class ImageProductController extends Controller
 {
     public function addImageProduct(Request $request)
     {
+        $this->authorize('checkemployee');
         $id = $request['id'];
 
         $file = $request->file('image_product');
@@ -27,6 +28,7 @@ class ImageProductController extends Controller
 
     public function destroyImageProduct(Request $request)
     {
+        $this->authorize('checkemployee');
         $imageProduct = ImageProduct::find($request['id']);
         File::delete($imageProduct->takeImage());
         $imageProduct->delete();
