+++ master/app/Http/Controllers/Admin/OrderController.php
@@ -11,6 +11,7 @@ class OrderController extends Controller
 {
     public function getWaitingCount()
     {
+        $this->authorize('checkemployee');
         $waiting_orders = count(Order::where('status', "waiting")->get());
         return response()->json(array(
             'waiting_orders' => $waiting_orders,
@@ -19,6 +20,7 @@ class OrderController extends Controller
 
     public function index()
     {
+        $this->authorize('checkemployee');
         $orders = new Order();
         $orders = $orders->join('users as customer', 'orders.customer_id', '=', 'customer.id')->leftJoin('users as employee', 'orders.employee_id', '=', 'employee.id');
         if (request('keyword')) {
@@ -34,11 +36,13 @@ class OrderController extends Controller
 
     public function read(Order $order)
     {
+        $this->authorize('checkemployee');
         return view('admins.orders.read', compact('order'));
     }
 
     public function changeStatus(Request $request)
     {
+        $this->authorize('checkemployee');
         $order = Order::find($request['id']);
         if($request['employee_id']) $order->employee_id = $request['employee_id'];
         $order->status = $request['status'];
@@ -49,11 +53,13 @@ class OrderController extends Controller
 
     public function create()
     {
+        $this->authorize('checkemployee');
         return view('admins.orders.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('checkemployee');
         $data = $request->all();
         Order::create($data);
         session()->flash('success', 'New Order Succesfully Created');
@@ -62,16 +68,19 @@ class OrderController extends Controller
 
     public function show(Order $order)
     {
+        $this->authorize('checkemployee');
         return view('admins.orders.show', compact('order'));
     }
 
     public function edit(Order $order)
     {
+        $this->authorize('checkemployee');
         return view('admins.orders.edit', compact('order'));
     }
 
     public function update(Request $request, Order $order)
     {
+        $this->authorize('checkowner');
         $attr = $request->all();
         $order->update($attr);
         session()->flash("success", "Order $order->id successfuly updated");
@@ -80,6 +89,7 @@ class OrderController extends Controller
 
     public function destroy(Order $order)
     {
+        $this->authorize('checkowner');
         $order->delete();
         session()->flash('success', "Order $order->id successfuly deleted");
         return redirect()->to(route('admins.orders.index'));
@@ -87,6 +97,7 @@ class OrderController extends Controller
 
     public function destroyNoReload(Request $request)
     {
+        $this->authorize('checkowner');
         $order = Order::find($request['id']);
         File::delete($order->takeImage());
         $order->delete();
